requirements.txt
setuptools>=65.0.0
Flask==2.3.3
numpy>=1.24.3,<2.0.0
psutil==5.9.5
gunicorn==21.2.0
cupy-cuda12x
nvidia-ml-py3
